[tests] bump test times; files must exist before uploading (#4696)
Context: https://build.azdo.io/3744248

Some of the MSBuild performance timings have started failing.

It seems like it is a change on the build machines, as I've not seen
anything in the logs pointing out a regression.  I merely bumped a
few of the times in `MSBuildDeviceIntegration.csv`.

`DeviceTest.CleanupTest()` would check for various expected output
files to add as "test attachments," e.g. the build log, and
`TestContext.AddTestAttachment()` would throw if the specified file
does not exist.

	TearDown : System.IO.FileNotFoundException : Test attachment file path could not be found.

Update `CleanupTest()` so that it checks that the file exists *before*
calling `TestContext.AddTestAttachment()`, avoiding the
`FileNotFoundException`.