[Xamarin.Android.Build.Tasks] fix .resx files for IDE builds (#4724)
Fixes: https://github.com/xamarin/xamarin-android/issues/4664
Fixes: https://feedback.devdiv.io/1030901

It turns out that building *inside* Visual Studio is causing `.resx`
files to stop working.  Satellite assemblies are missing from the
`.apk` and/or were not deployed with Fast Deployment.

In 65a917a9, I removed a direct call to `<ResolveAssemblyReference/>`
for performance reasons. This appears to have caused this problem.

This is due to:

  * https://github.com/microsoft/msbuild/blob/dc485bce3427e9d2b020ce61c2400e7b5a76062c/src/Tasks/Microsoft.Common.CurrentVersion.targets#L2101
  * https://github.com/microsoft/msbuild/blob/dc485bce3427e9d2b020ce61c2400e7b5a76062c/src/Tasks/Microsoft.Common.CurrentVersion.targets#L2059

`<ResolveAssemblyReference/>` by default only finds satellite
assemblies when `$(BuildingProject)` is set.  This does *not* happen
when `$(BuildingInsideVisualStudio)` is `true` during design-time
builds.

Thus, inside the IDE:

  * `Build` runs, `$(BuildingProject)` is `true` and everything works.
  * `Install` runs, `$(BuildingProject)` is `false`.  This is the part
    where everything falls apart.

I could reproduce this in a test by running `Build` then
`SignAndroidPackage` in two steps.  Our MSBuild tests already set
`$(BuildingInsideVisualStudio)` to `true` by default.

The fix for this is to make `SignAndroidPackage` depend on
`BuildOnlySettings`:

  * https://github.com/microsoft/msbuild/blob/dc485bce3427e9d2b020ce61c2400e7b5a76062c/src/Tasks/Microsoft.Common.CurrentVersion.targets#L1081-L1086

This makes `$(BuildingProject)` always `true` if `SignAndroidPackage`
or `Install` runs in the IDE.