[xaprepare] Improve Linux OS detection
Xamarin.Android supports detecting a handful of Linux distributions
directly (Ubuntu, Linux Mint, Debian and Arch) but the support can be
improved so that derivatives of those distributions can be detected
properly without having direct support for them.  To do it, this commit
adds support to detect derivatives of the above 4 distributions by way
of looking at the `/etc/os-release` file which contains all the
information we need.