[tests] CodeBehindTests explicitly set $(TFV)
The following import does not work when the tests run:

    <Import
        Condition="Exists('..\..\..\Configuration.props')"
        Project="..\..\..\Configuration.props"
    />

`CodeBehindBuildTests.csproj` is copied to a directory, where it would
need to use this path:

    ..\..\..\..\..\..\..\Configuration.props

But the first path is correct, for when the project is built in-place
in `Xamarin.Android-Tests.sln`.

This path is getting too long... I think it is simpler to set
`$(TargetFrameworkVersion)` in the test, as we can use
`builder.LatestTargetFrameworkVersion ()`.