[tests] mark F# tests NonParallelizable
Context: https://build.azdo.io/3775041

We've been seeing the F# tests fail with errors such as:

    NuGet.targets(124,5): error : Could not find file "/Users/runner/.local/share/NuGet/v3-cache/1ca707a4d90792ce8e42453d4e350886a0fdaa4d$ps:_api.nuget.org_v3_index.json/nupkg_fsharp.core.4.7.1.dat"

These are likely due to `FSharp.Core` being restored in parallel.

I added `[NonParallelizable]` to the F# tests, which has fixed this
problem for certain tests in the past.

I also cleaned up the `BuildBasicApplicationFSharp` test to use a
parameter for `Debug` vs `Release`.