[generator] Don't invalidate interface if static method is invalidated (#660)
Fixes: https://github.com/xamarin/java.interop/issues/588

If we cannot bind a method on an interface we mark the interface as
unbindable and omit the interface from the generated code, and
invalidate anything that depends on the interface.  This is because
an interface with a missing method cannot be implemented in C#.

However, if the method is `static` then we *can* remove the method
while keeping the interface, as the user is not required to implement
the static method.

Update `GenBase.OnValidate()` so that invalid static methods don't
invalidate their declaring interface.

Note we already properly handled Default Interface Methods, this
adds support for `static` methods as well.