[compression] Store assembly information with unique key
Context: d236af54538ef1fe62d0125798cdde78e46dcd8e

d236af54 introduced assembly compression and one part of the process is
to gather information about the size of all the assemblies to be
packaged with the application, in order to generate native assembly code
which pre-allocates memory for them.  To do that, the native runtime must
be sure that the the assembly being currently decompressed will fit in
the pre-allocated memory area and thus it stores the original file size
both in the compressed assembly header and in the generated native
assembler source (compiled into `libxamarin-app.so`).  If the sizes
mismatch, we will see an error in the logcat similar to:

    F/monodroid-assembly(11348): Compressed assembly 'mscorlib.dll' is larger than when the application was built (expected at most 146432, got 2043392). Assemblies don't grow just like that!

In addition to the assembly size, the compressed assembly header
contains also an index into an array of assembly descriptors which is
placed in the generated native assembler source code.  The
index **must** point to the correct descriptor or bad things happen.
This is the reason for the particular error shown above, coming from a
test app (here)[0].

During build time, the information about assembly sizes is carried
between the involved tasks in an object registered with MSBuild using
its `BuildEngine4.RegisterTaskObject` API.  The API uses a key to access
the registered object which, in d236af54, was defined to be a constant
string, shared across all the builds.  However, it turns out
that *somehow* if two or more projects store the registered object with
the same key the end result is that the cache contains entries from more
than one object (how it happened is unclear, it should be "impossible")
which results in the descriptor indexes in compressed assembly header to
point to incorrect locations.

In order to fix the issue, this commit makes two changes:

 * Make sure the assembly info object is unregistered by the `BuildApk`.
   This ensures that "merging" of two dictionaries (however/wherever it
   happened) will not take place.
 * Make the key to the registered object unique by addition of the
   project GUID (from the `$(ProjectGuid)` MSBuild property).
   This ensures correct operation during builds which take advantage of
   MSBuild's parallel build feature.

[0]: https://github.com/xamarin/QualityAssurance/tree/master/Samples/android/Bug15162

Co-authored-by: Jonathan Peppers <jonathan.peppers@gmail.com>