[Xamarin.Android.Build.Tasks] add $(AndroidSupportedAbis) to build.props
Fixes: https://github.com/xamarin/xamarin-android/issues/4790

If you change `$(AndroidSupportedAbis)` in the IDE and perform an
incremental build, you can hit:

    [i686-linux-android-as.EXE stderr] Assembler messages:
    [i686-linux-android-as.EXE stderr] Error: can't open typemaps.x86.s for reading: No such file or directory
    Xamarin.Android.Common.targets(2245,3): error XA3006: Could not compile native assembly file: typemaps.x86.s

I was able to reproduce the issue in a test by using the `.csproj`:

    <AndroidSupportedAbis>armeabi-v7a</AndroidSupportedAbis>

Then command-line pass a different ABI, such as:

    -p:AndroidSupportedAbis=x86

`$(AndroidSupportedAbis)` needs to be listed in `build.props`, so that
changes to this value will trigger all MSBuild targets to run again.
In particular `_GenerateJavaStubs` and `_GeneratePackageManagerJava`
need to run in this case.