[One .NET] updated implementation for 'dotnet publish'
Currently, `dotnet publish` is incomplete uses the built-in behavior
in dotnet/sdk that copies all the managed assemblies to a `publish`
directory. Ideally an `.apk` or `.aab` should be the output instead?

For example:

    > dotnet publish foo.android.csproj

This should copy any `.apk` files to `$(PublishDir)` and nothing else.

To achieve this, we need to declare our own minimal `Publish` target
that depends on:

    Build;
    PrepareForPublish;
    CopyFilesToPublishDirectory;

We don't want *our* `Publish` target to run things like `ILLink`, AOT,
etc. because our `Build` target will be in charge of running those
build steps.

The current implementation in `Microsoft.Android.Sdk.Publish.targets`
is based on the `_CopyResolvedFilesToPublishPreserveNewest` target in
dotnet/sdk:

https://github.com/dotnet/sdk/blob/22fc243c1f6c9455a56bda014f1c32b6e1af18fc/src/Tasks/Microsoft.NET.Build.Tasks/targets/Microsoft.NET.Publish.targets#L218-L240

The only hangup was making sure:

    > dotnet publish foo.android.csproj -p:AndroidPackageFormat=apk
    > dotnet publish foo.android.csproj -p:AndroidPackageFormat=aab

After the second command, the `_IncrementalCleanPublishDirectory`
MSBuild target needs to actually cleanup the previous `.apk` files.

https://github.com/dotnet/sdk/blob/22fc243c1f6c9455a56bda014f1c32b6e1af18fc/src/Tasks/Microsoft.NET.Build.Tasks/targets/Microsoft.NET.Publish.targets#L137-L157

To make this work, we need to fixup `@(ResolvedFileToPublish)` before
the `CopyFilesToPublishDirectory` MSBuild target runs. It also needs
the `%(RelativePath)` metadata to be filled out with the expected path
after `$(PublishDir)`.

Down the road we can extend the `Publish` target for other scenarios.

I added a test changing `$(AndroidPackageFormat)` and verifying the
correct files land in `$(PublishDir)`. I also updated `DotNetCLI`, so
you can pass global MSBuild parameters.