Install apkdiff as dotnet tool (#4799)
instead of building it.

As the [apkdiff](https://github.com/radekdoulik/apkdiff) is now distributed as
[nuget](https://www.nuget.org/packages/apkdiff/) and can be installed as dotnet tool, it doesn't
make sense to have it in the XA repo anymore.

The CI yaml was updated to install and use the `apkdiff` through the `dotnet` tool. The sources were
removed from the tree and `xaprepare` updated.