[Xamarin.Android.Build.Tasks] fix IDE builds missing satellite assemblies in apps (#4813)
Fixes: https://github.com/xamarin/xamarin-android/issues/4805
Context: https://github.com/xamarin/xamarin-android/issues/4664

There are still reports of `.resx` files not working inside IDEs...
65a917a9 was the original cause of this problem.

Unfortunately, the original problem in #4664 still occurs if the
`.resx` files are in the Xamarin.Android application project. The
original fix in d298fdaa, only solved the problem for missing
satellite assemblies in referenced projects.

Two missing targets need to run for `SignAndroidPackage` in IDE builds:

    PrepareResources;
    CreateSatelliteAssemblies;

These targets fill out the
`@(IntermediateSatelliteAssembliesWithTargetPath)` item group, that
contains all satellite assemblies for the current project.

In IDE builds, this item group was empty and `<BuildApk/>` was not
adding satellite assemblies to the `.apk` file at all.

I was able to reproduce this problem pretty easily with a test. I
reworked the `MissingSatelliteAssembly` test to have a version for
libraries and apps.

I also cleaned the test up a bit, in general, using less hardcoded
strings.