[Xamarin.Android.Build.Tasks] @(LinkDescription) should be in Inputs for the linker (#4816)
Fixes: https://developercommunity.visualstudio.com/content/problem/1061307/changes-not-applied-on-build-after-linker-configur.html

Changes to a `@(LinkDescription)` file did not trigger the
`_LinkAssembliesShrink` MSBuild target to run again.

The only way to see changes from this file was to `Rebuild` every time.

Adding a test for this scenario is also nice, as it didn't seem like
we had an MSBuild test verifying that `@(LinkDescription)` works in
general.