[Xamarin.Android.Build.Tasks] Assembly compression info fixes (#4792)
Context: d236af54538ef1fe62d0125798cdde78e46dcd8e

Bumps to xamarin/monodroid/master@4a0be740

Changes: https://github.com/xamarin/monodroid/compare/0d0e4527f142e73bdff9fa1b63349bb0a1631900...6d05c4929db1ba12ba3a04174938256e435dbbd4

  * xamarin/monodroid@4a0be740: Pass a couple parameters to `<BuildApk>`

Commit d236af54 introduced assembly compression and one part of the
process is to gather information about the size of all the assemblies
to be packaged with the application, in order to generate native
assembly code which pre-allocates memory for them.  To do that, the
native runtime must be sure that the the assembly being currently
decompressed will fit in the pre-allocated memory area, and thus it
stores the original file size both in the compressed assembly header
and in the generated native assembler source, compiled into
`libxamarin-app.so`.

If the sizes mismatch, we will see an error in `adb logcat`, e.g.:

	F/monodroid-assembly(11348): Compressed assembly 'mscorlib.dll' is larger than when the application was built (expected at most 146432, got 2043392). Assemblies don't grow just like that!

In addition to the assembly size, the compressed assembly header also
contains an index into an array of assembly descriptors which is placed
in the generated native assembler source code.  The index **must**
point to the correct descriptor or bad things happen.  This is the
reason for the particular error shown above, via [a test app][0].

During build time, the information about assembly sizes is carried
between the involved tasks in an object registered with MSBuild using
[`IBuildEngine4.RegisterTaskObject()`][1] API.   The API uses a key to
access the registered object which, in d236af54, was defined to be a
constant string, shared across all the builds.  However, it turns out
that *somehow* if two or more projects store the registered object with
the same key the end result is that the cache contains entries from
more than one object, which results in the descriptor indexes in the
compressed assembly header to point to incorrect locations.
(How this happened is unclear, it should be "impossible".)

In order to fix the issue, this make two changes:

  * Make sure the assembly info object is unregistered by the
    `<BuildApk/>` task.  This ensures that "merging" of two
    dictionaries (however/wherever it happened) will not take place.

  * Make the key to the registered object unique by the addition of the
    project GUID (from the `$(ProjectGuid)` MSBuild property).
    This ensures correct operation during builds which take advantage
    of MSBuild's parallel build feature.

[0]: https://github.com/xamarin/QualityAssurance/tree/master/Samples/android/Bug15162
[1]: https://docs.microsoft.com/en-us/dotnet/api/microsoft.build.framework.ibuildengine4.registertaskobject?view=netframework-4.8

Co-authored-by: Jonathan Peppers <jonathan.peppers@gmail.com>