[Xamarin.Android.Build.Tasks] use MSBuild items for ProGuard rules (#4825)
Context: https://github.com/xamarin/monodroid/commit/e82633e9eaf36b2ac3530a681c86d6926bf7f8c8
Context: https://github.com/xamarin/Java.Interop/commit/1de5501b2d49df3d8d3b187df026d611996a67da
Context: https://github.com/microsoft/msbuild/issues/1024#issuecomment-248669947

When using `dotnet build`, usage of R8 or ProGuard could encounter
build warnings on macOS such as:

	warning XA4304: ProGuard configuration file '/Users/runner/Library/Android/sdk//tools\proguard\proguard-android.txt' was not found.

Reviewing how this worked, the `$(ProguardConfigFiles)` MSBuild
property:

	<ProguardConfigFiles Condition="'$(ProguardConfigFiles)' == '' And '$(AndroidLinkTool)' != ''">
	    {sdk.dir}tools\proguard\proguard-android.txt;
	    {intermediate.common.xamarin};
	    {intermediate.references};
	    {intermediate.application};
	    @(ProguardConfiguration);
	</ProguardConfigFiles>

was using string replacements such as:

	var configs = ProguardConfigurationFiles
	    .Replace ("{sdk.dir}", AndroidSdkDirectory + Path.DirectorySeparatorChar)
	    .Replace ("{intermediate.common.xamarin}", ProguardCommonXamarinConfiguration)
	    .Replace ("{intermediate.references}", ProguardGeneratedReferenceConfiguration)
	    .Replace ("{intermediate.application}", ProguardGeneratedApplicationConfiguration)
	    .Replace ("{project}", string.Empty) // current directory anyways.
	    .Split (';')
	    .Select (s => s.Trim ())
	    .Where (s => !string.IsNullOrWhiteSpace (s));

The above set of `.Replace()` calls would result in a path such as:

	/Users/runner/Library/Android/sdk//tools\proguard\proguard-android.txt

Note the `\` after `tools`.

When building on Linux or macOS under Mono+msbuild, the `\` is replaced
with a `/`, resulting in a valid path.

However, `dotnet build` does not replace `\` with `/` in as many places
as Mono's msbuild, and the above path *retains the `\`* in a subsequent
`File.Exists()` call -- along with passing via `proguard --pg-conf` --
and as `\` is a Perfectly Valid Filename Character on Unix, the result
is a filename which *does not exist*.

At minimum, we need to ensure that `\` is replaced with `/` on Unixy
platforms in order for things to work under `dotnet build`.  This is
done for `<ItemGroup/>`s, but *not* for MSBuild properties.

But why are we even replacing strings like this?

Reading the ProGuard documentation, `{sdk.dir}` and `{project}` do not
seem to be a placeholder we need to support.  There is [a mention][0]
of `${sdk.dir}` (with a dollar sign), but this would produce:

	$/Users/runner/Library/Android/sdk//tools\proguard\proguard-android.txt

This leads me to believe Xamarin.Android invented its own string
templating for this instead of embracing MSBuild.  It was a long time
ago (~2014), so xbuild support was also likely a factor then.

To clean things up and fix .NET 5 at the same time:

 1. Use `@(_ProguardConfiguration)` instead of `$(ProguardConfigFiles)`.
 2. List the files as you would in MSBuild item groups normally.
 3. Use `string[]` as the value type in the `<Proguard/>` and `<R8/>`
    MSBuild tasks.
 4. Continue to support the `$(ProguardConfigFiles)` property for
    backwards compatibility. It can remain an escape hatch if someone
    wants to replace *every* file, though using the
    [`Remove` attribute][1] should be preferred.

Other changes:

  * Updated the `BuildHasNoWarnings()` test, so it will run under
    `dotnet build`.  Don't set `$(MonoSymbolArchive)` to `true` when
    running under `dotnet build`.

  * The `$(AndroidSdkDirectory)` property is no longer needed in the
    `<Proguard/>` and `<R8/>` MSBuild tasks.

  * Removed `System.Linq` usage.

[0]: https://www.guardsquare.com/en/products/proguard/manual/examples
[1]: https://docs.microsoft.com/en-us/visualstudio/msbuild/msbuild-items?view=vs-2019#BKMK_RemoveAttribute